+++ src/__tests__/edit.test.js
@@ -1,0 +1,19 @@
+import { editTask } from "../crud,js";
+
+jest.mock('../localStorage');
+
+describe('Edit a task', () => {
+  test('Happy Path Example 1', () => {
+    document.body.innerHTML = `<ul id="list">
+      <li class="editing">
+        <div class  = "d-flex-between">
+        <input type="checkbox" class="checks" id="2" value="2">
+        <input type="text" value="task2" class="edit-task" disabled>
+        </div> 
+        <button class="delete-button task-buttons d-none">
+          <i class="bi bi-trash"></i>
+        </button>
+      </li>
+    </ul>`;
+  });
+});
